feat(Network): Moved Explore page items to Network page
