+++ apps/web/src/app/[lang]/network/page.tsx
@@ -1,4 +1,9 @@
 import { Locale } from '@hypha-platform/i18n';
+import { Container } from '@hypha-platform/ui';
+import { Text } from '@radix-ui/themes';
+import { SpaceGroupSlider, Search } from '@hypha-platform/epics';
+import { readAllSpaces } from '../../actions/space';
+import { getDhoPathAgreements } from '../dho/[id]/agreements/constants';
 
 type PageProps = {
   params: Promise<{ lang: Locale; id: string }>;
@@ -9,5 +14,19 @@ export default async function Index(props: PageProps) {
 
   const { lang } = params;
 
-  return <div>NetworkPage</div>;
+  const getHref = (id: string) => {
+    return getDhoPathAgreements(lang, id);
+  };
+
+  const spaces = await readAllSpaces();
+
+  return (
+    <Container>
+      <Text className="text-9 text-center flex mb-8">
+        Explore hundreds of Spaces in the Hypha Network
+      </Text>
+      <Search />
+      <SpaceGroupSlider spaces={spaces} type="Hypha" getHref={getHref} />
+    </Container>
+  );
 }
